requirements.txt
mcstatus
discord.py